refactor: centralize all tool definitions in coreTools.ts
