+++ packages/core/src/tools/shell.ts
@@ -43,7 +43,7 @@ import {
 } from '../utils/shell-utils.js';
 import { SHELL_TOOL_NAME } from './tool-names.js';
 import type { MessageBus } from '../confirmation-bus/message-bus.js';
-import { SHELL_DEFINITION } from './definitions/coreTools.js';
+import { getShellDefinition } from './definitions/coreTools.js';
 import { resolveToolDeclaration } from './definitions/resolver.js';
 
 export const OUTPUT_UPDATE_INTERVAL_MS = 1000;
@@ -453,41 +453,6 @@ export class ShellToolInvocation extends BaseToolInvocation<
   }
 }
 
-export function getShellToolDescription(
-  enableInteractiveShell: boolean,
-): string {
-  const returnedInfo = `
-
-      The following information is returned:
-
-      Output: Combined stdout/stderr. Can be \`(empty)\` or partial on error and for any unwaited background processes.
-      Exit Code: Only included if non-zero (command failed).
-      Error: Only included if a process-level error occurred (e.g., spawn failure).
-      Signal: Only included if process was terminated by a signal.
-      Background PIDs: Only included if background processes were started.
-      Process Group PGID: Only included if available.`;
-
-  if (os.platform() === 'win32') {
-    const backgroundInstructions = enableInteractiveShell
-      ? 'To run a command in the background, set the `is_background` parameter to true. Do NOT use PowerShell background constructs.'
-      : 'Command can start background processes using PowerShell constructs such as `Start-Process -NoNewWindow` or `Start-Job`.';
-    return `This tool executes a given shell command as \`powershell.exe -NoProfile -Command <command>\`. ${backgroundInstructions}${returnedInfo}`;
-  } else {
-    const backgroundInstructions = enableInteractiveShell
-      ? 'To run a command in the background, set the `is_background` parameter to true. Do NOT use `&` to background commands.'
-      : 'Command can start background processes using `&`.';
-    return `This tool executes a given shell command as \`bash -c <command>\`. ${backgroundInstructions} Command is executed as a subprocess that leads its own process group. Command process group can be terminated as \`kill -- -PGID\` or signaled as \`kill -s SIGNAL -- -PGID\`.${returnedInfo}`;
-  }
-}
-
-export function getCommandDescription(): string {
-  if (os.platform() === 'win32') {
-    return 'Exact command to execute as `powershell.exe -NoProfile -Command <command>`';
-  } else {
-    return 'Exact bash command to execute as `bash -c <command>`';
-  }
-}
-
 export class ShellTool extends BaseDeclarativeTool<
   ShellToolParams,
   ToolResult
@@ -501,12 +466,13 @@ export class ShellTool extends BaseDeclarativeTool<
     void initializeShellParsers().catch(() => {
       // Errors are surfaced when parsing commands.
     });
+    const definition = getShellDefinition(config.getEnableInteractiveShell());
     super(
       ShellTool.Name,
       'Shell',
-      getShellToolDescription(config.getEnableInteractiveShell()),
+      definition.base.description!,
       Kind.Execute,
-      SHELL_DEFINITION.base.parameters!,
+      definition.base.parameters!,
       messageBus,
       false, // output is not markdown
       true, // output can be updated
@@ -546,18 +512,13 @@ export class ShellTool extends BaseDeclarativeTool<
   }
 
   override getSchema(modelId?: string) {
+    const definition = getShellDefinition(
+      this.config.getEnableInteractiveShell(),
+    );
     if (!modelId) {
       return super.getSchema();
     }
 
-    const declaration = resolveToolDeclaration(SHELL_DEFINITION, modelId);
-    const schema = super.getSchema();
-
-    // We use the resolved declaration but preserve the dynamic platform-specific description
-    // from the tool instance to ensure behavior remains 100% identical for now.
-    return {
-      ...declaration,
-      description: schema.description,
-    };
+    return resolveToolDeclaration(definition, modelId);
   }
 }
